Add Search Insert Position project and implementation
Updated the solution to include a new project titled "Search Insert Position (O)" with the source file `Optimal.cpp`. Modified the solution file to add build configurations for the new project. Implemented a `Solution` class with a `searchInsert` method to find the index of a target in a vector. Created project and filters files to define project settings and organize source files.