Add Jump Game project to Data Structures & Algorithms
This commit introduces a new project named "Jump Game" to the solution "Data Structures & Algorithms". It includes updates to the solution file to accommodate the new project with configurations for Debug and Release modes across Win32 and x64 platforms.

Additionally, a project file (`Jump Game.vcxproj`) and a filters file (`Jump Game.vcxproj.filters`) were created to define project settings and organize files. A new source file (`Optimal.cpp`) was added, implementing a class `Solution` with a method `canJump` to determine if one can reach the last index of an array based on specified jump lengths.